+++ clients/cli/src/utils.rs
@@ -10,11 +10,9 @@ use crate::error::CliResult;
 use crate::utils::SupportedEditors::{Code, Emacs, Nano, Sublime, Vim};
 use crate::{err, err_extra, err_unexpected};
 use hotwatch::{Event, Hotwatch};
+use lockbook_core::pure_functions::drawing::SupportedImageFormats;
+use lockbook_core::pure_functions::drawing::SupportedImageFormats::*;
 use lockbook_core::service::db_state_service::State;
-use lockbook_core::service::drawing_service::SupportedImageFormats;
-use lockbook_core::service::drawing_service::SupportedImageFormats::{
-    Bmp, Farbfeld, Jpeg, Png, Pnm, Tga,
-};
 use lockbook_models::account::Account;
 use uuid::Uuid;
