+++ core/src/external_interface/c_interface.rs
@@ -9,10 +9,9 @@ use uuid::Uuid;
 use lockbook_models::file_metadata::FileType;
 
 use crate::external_interface::json_interface::translate;
-use crate::get_all_error_variants;
 use crate::model::state::Config;
-use crate::service::drawing_service::SupportedImageFormats;
 use crate::service::path_service::{filter_from_str, Filter};
+use crate::{get_all_error_variants, SupportedImageFormats};
 
 fn c_string(value: String) -> *const c_char {
     CString::new(value)
