+++ core/src/pure_functions/mod.rs
@@ -1,1 +1,2 @@
+pub mod drawing;
 pub mod files;
